+++ Rdmp.Core/Providers/DataExportChildProvider.cs
@@ -356,9 +356,6 @@ public class DataExportChildProvider : CatalogueChildProvider
     private void AddChildren(ProjectCohortsNode projectCohortsNode, DescendancyList descendancy)
     {
         var children = new HashSet<object>();
-        var projectCiCsNode = new ProjectCohortIdentificationConfigurationAssociationsNode(projectCohortsNode.Project);
-        children.Add(projectCiCsNode);
-        AddChildren(projectCiCsNode, descendancy.Add(projectCiCsNode));
 
         var savedCohortsNode = new ProjectSavedCohortsNode(projectCohortsNode.Project);
         children.Add(savedCohortsNode);
@@ -400,38 +397,6 @@ public class DataExportChildProvider : CatalogueChildProvider
         {
             AddChildren(cohortSourceUsedByProjectNode, descendancy.Add(cohortSourceUsedByProjectNode));
             children.Add(cohortSourceUsedByProjectNode);
-        }
-
-        AddToDictionaries(children, descendancy);
-    }
-
-    private void AddChildren(ProjectCohortIdentificationConfigurationAssociationsNode projectCiCsNode,
-        DescendancyList descendancy)
-    {
-        //add the associations
-        var children = new HashSet<object>();
-        foreach (var association in AllProjectAssociatedCics.Where(assoc =>
-                     assoc.Project_ID == projectCiCsNode.Project.ID))
-        {
-            var matchingCic = AllCohortIdentificationConfigurations.SingleOrDefault(cic =>
-                cic.ID == association.CohortIdentificationConfiguration_ID) ?? AllTemplateCohortIdentificationConfigurations.SingleOrDefault(cic =>
-                cic.ID == association.CohortIdentificationConfiguration_ID);
-
-            if (matchingCic == null)
-            {
-                _errorsCheckNotifier.OnCheckPerformed(
-                    new CheckEventArgs(
-                        $"Failed to find Associated Cohort Identification Configuration with ID {association.CohortIdentificationConfiguration_ID} which was supposed to be associated with {association.Project}",
-                        CheckResult
-                            .Fail)); //inject knowledge of what the cic is so it doesn't have to be fetched during ToString
-            }
-            else
-            {
-                association.InjectKnown(matchingCic);
-
-                //document that it is a child of the project cics node
-                children.Add(association);
-            }
         }
 
         AddToDictionaries(children, descendancy);